+++ juniper/src/executor_tests/enums.rs
@@ -1,4 +1,4 @@
-use juniper_codegen::GraphQLEnumInternal as GraphQLEnum;
+use juniper_codegen::EnumInternal as GraphQLEnum;
 
 use crate::ast::InputValue;
 use crate::executor::Variables;